# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/leads": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Список",
                "operationId": "a343b1c7e5e7371cc2c8e8d3daf86770",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "company": {
                                                        "type": "string",
                                                        "example": "Some company"
                                                    },
                                                    "fullname": {
                                                        "type": "string",
                                                        "example": "Данило Галицький"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+380964545678"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "example@mail.com"
                                                    },
                                                    "comment": {
                                                        "type": "string",
                                                        "example": "Some comment"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "новий"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Создание",
                "operationId": "6bbdb9ce33bed941e06daceafdc6b2dd",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "company": {
                                                "type": "string",
                                                "example": "Some company"
                                            },
                                            "fullname": {
                                                "type": "string",
                                                "example": "Данило Галицький"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "+380964545678"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "example@mail.com"
                                            },
                                            "comment": {
                                                "type": "string",
                                                "example": "Some comment"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "новий"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "company": {
                                                    "type": "string",
                                                    "example": "Some company"
                                                },
                                                "fullname": {
                                                    "type": "string",
                                                    "example": "Данило Галицький"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+380964545678"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "example@mail.com"
                                                },
                                                "comment": {
                                                    "type": "string",
                                                    "example": "Some comment"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "новий"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/leads/{lead}": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Единичная запись",
                "operationId": "067a8fa82bd6b082e2fe3098dfe1d7e7",
                "parameters": [
                    {
                        "name": "lead",
                        "in": "path",
                        "description": "ID лида",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "company": {
                                                    "type": "string",
                                                    "example": "Some company"
                                                },
                                                "fullname": {
                                                    "type": "string",
                                                    "example": "Данило Галицький"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+380964545678"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "example@mail.com"
                                                },
                                                "comment": {
                                                    "type": "string",
                                                    "example": "Some comment"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "новий"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Leads"
                ],
                "summary": "Удаление",
                "operationId": "57b4679aaf55e26380e66e0aed5ab498",
                "parameters": [
                    {
                        "name": "lead",
                        "in": "path",
                        "description": "ID лида",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Leads"
                ],
                "summary": "Обновление",
                "operationId": "92d3f2996ffa990d16e4b877681dc356",
                "parameters": [
                    {
                        "name": "lead",
                        "in": "path",
                        "description": "ID лида",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "company": {
                                                "type": "string",
                                                "example": "Some company"
                                            },
                                            "fullname": {
                                                "type": "string",
                                                "example": "Данило Галицький"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "+380964545678"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "example@mail.com"
                                            },
                                            "comment": {
                                                "type": "string",
                                                "example": "Some comment"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "новий"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "company": {
                                                    "type": "string",
                                                    "example": "Some company"
                                                },
                                                "fullname": {
                                                    "type": "string",
                                                    "example": "Данило Галицький"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+380964545678"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "example@mail.com"
                                                },
                                                "comment": {
                                                    "type": "string",
                                                    "example": "Some comment"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "новий"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}